Add support for downscaler/force-uptime
Look for any pods that have downscaler/force-uptime annotation set to
true. If at least one is present, force the whole cluster back into
uptime until the pods are terminated.

Change the default polling interval to 30 seconds so users don't need to
wait too long for the scale-up to finish.